CFY-6648 Move events V3 test cases to a separate file
